[codex] split main entrypoint into modules (#105)
* split main entrypoint into modules

* fix(ui): resolve NameErrors causing TUI crashes

- Imported `cast` in `ui_textual.py` to fix a crash when rendering the configuration preview.

- Replaced separate source count variables with a `counts` dictionary to fix a `NameError` and properly normalize job source strings.

* Update src/gengowatcher/cli.py


* fix: apply CodeRabbit auto-fixes

Fixed 3 file(s) based on 4 unresolved review comments.


* fix: apply CodeRabbit auto-fixes

Fixed 2 file(s) based on 2 unresolved review comments.


---------
